requirements.txt
ollama==0.4.7
discord~=2.3.2
discord.py~=2.5.0
python-dotenv~=1.0.1